tools/diffusiondrive/server/requirements.txt
huggingface-hub>=0.20.3
numpy>=1.24
opencv-python>=4.8.0
pyyaml>=6.0
pyzmq>=25.1
tqdm>=4.66
